gamelib-backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
bcrypt
python-jose[cryptography]
python-multipart
google-auth
requests
